optimize token defaults: aggressive pruning, long cache, tighter compaction
- Context pruning: lower softTrimRatio (0.3->0.25), hardClearRatio (0.5->0.4),
  minPrunableToolChars (50k->30k), softTrim maxChars (4k->3k) to reduce
  stale tool result bloat and cut input tokens per request
- Cache retention: default to "long" (1h) instead of "short" (5m) for
  Anthropic API key auth, reducing cache write costs
- Compaction chunks: lower BASE_CHUNK_RATIO (0.4->0.35) and MIN_CHUNK_RATIO
  (0.15->0.12) for smaller, more focused summarization passes
- Reserve tokens floor: lower from 20k to 16k to delay unnecessary
  compaction summarization API calls

+++ src/config/config.pruning-defaults.test.ts
@@ -101,7 +101,7 @@ describe("config pruning defaults", () => {
       expect(cfg.agents?.defaults?.heartbeat?.every).toBe("30m");
       expect(
         cfg.agents?.defaults?.models?.["anthropic/claude-opus-4-5"]?.params?.cacheRetention,
-      ).toBe("short");
+      ).toBe("long");
     });
   });
